pyproject.toml
[project]
name = "fastapi-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.136.1",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
